Move ConnectionTestWithCutltureTH back to SqlConnectionBasicTests. Replaced [ConditionalFact] with just [Fact] Added catch { throw; } back.
